+++ common/src/main/java/com/technoserv/db/service/objectmodel/impl/BioTemplateVersionServiceImpl.java
@@ -12,7 +12,7 @@ import org.springframework.stereotype.Service;
  * Created by sergey on 23.11.2016.
  */
 @Service
-public class BioTemplateVersionServiceImpl extends AbstractService<Integer,BioTemplateVersion,BioTemplateVersionDao> implements BioTemplateVersionService {
+public class BioTemplateVersionServiceImpl extends AbstractService<Long,BioTemplateVersion,BioTemplateVersionDao> implements BioTemplateVersionService {
     @Override
     @Autowired
     @Qualifier("bioTemplateVersionDao")
